% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_empty_matrix_and_fill_it.r
\name{create_empty_matrix_and_fill_it}
\alias{create_empty_matrix_and_fill_it}
\title{This function creates an empty matrix with designated number of rows and columns, filled by "fill"}
\usage{
create_empty_matrix_and_fill_it(rnum, colnum, fill = 0)
}
\arguments{
\item{rnum}{Number of rows.}

\item{colnum}{Number of columns}

\item{fill}{Variable to fill empty fields with.}
}
\value{
Matrix, filled with fill variable.
}
\description{
This function creates an empty matrix with designated number of rows and columns, filled by "fill"
}
