% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hash_link_for_multiple_values_per_key.r
\name{create_hash_link_for_multiple_values_per_key}
\alias{create_hash_link_for_multiple_values_per_key}
\title{This function creates hash link for several different values per single unique key. Basically appends value with second instance it encounters etc.}
\usage{
create_hash_link_for_multiple_values_per_key(
  keys,
  values,
  input_hash = NA,
  make_unique = TRUE,
  do_sort = TRUE
)
}
\arguments{
\item{keys}{Vector of keys, same as you would pass to .set for hash.}

\item{values}{Vector of values, same as you would pass to .set for hash.}

\item{input_hash}{Hash to add values to. If none is provided it is created within script (default)}

\item{make_unique}{Do you only want unique values in hash returnes?}

\item{do_sort}{Do you want unique values in hash returnes sorted?}
}
\value{
Hash with appended values, wich returns vectors of values as input.
}
\description{
This function creates hash link for several different values per single unique key. Basically appends value with second instance it encounters etc.
}
