% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_matrix_to_an_edge_pair.r
\name{distance_matrix_to_an_edge_pair}
\alias{distance_matrix_to_an_edge_pair}
\title{This function converts symmetrical distance matrixes (such as matrixes, obtained by as.matrix from an output of dist function) into data frame of edges (with distance).}
\usage{
distance_matrix_to_an_edge_pair(x)
}
\arguments{
\item{x}{Symmetrical distance matrix.}
}
\value{
Data frame of edge pairs (with distance).
}
\description{
This function converts symmetrical distance matrixes (such as matrixes, obtained by as.matrix from an output of dist function) into data frame of edges (with distance).
}
