% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_string_merger.r
\name{dna_string_merger}
\alias{dna_string_merger}
\title{This function is for working with vcfs with multiple alternative alleles. It collapses DNA string from alt allele field of variant annotation file to get merged string with comma delimited alt alleles.}
\usage{
dna_string_merger(dnastring, collapse = ", ")
}
\arguments{
\item{dnastring}{A vector of alternative alleles.}

\item{collapse}{A delimiter to put stuff in.}
}
\value{
A string of with collapse delimited alt alleles.
}
\description{
This function is for working with vcfs with multiple alternative alleles. It collapses DNA string from alt allele field of variant annotation file to get merged string with comma delimited alt alleles.
}
