% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_limma.r
\name{do_limma}
\alias{do_limma}
\title{This will be a test run for adding roxygen comment to files
This is wrapper to run limma. Warning - it requires specifically formatted local ensembl annotation to run correctly}
\usage{
do_limma(
  count_matrix,
  design_matrix,
  output_file_result,
  output_file_degs,
  local_ensembl_annotation,
  contrast_num = 2,
  threshold_fc = 1.5,
  threshold_pval = 0.05,
  method_string = "hommel",
  output_plot_path = NA,
  pre_filtration = TRUE,
  do_fc_filtration = TRUE
)
}
\arguments{
\item{count_matrix}{Numeric matrix. A matrix with genes in rows and samples in columns}

\item{design_matrix}{Numeric matrix. A design matrix, with smaples in row matching columns of count and and variables in columns. Should be generated through model.matrix from stats}

\item{output_file_result}{A path to a file. FC and p.val for chosen contrast for each expressed gene are going to be written in the file with that path}

\item{output_file_degs}{A path to a file. FC and p.val for only degs (if any) for each expressed gene are going to be written in the file with that path}

\item{local_ensembl_annotation}{A path to a file. File with specially prepared local annotations TO ADD IN THE FUTURE - SCRIPT FOR CREATING THOSE}

\item{contrast_num}{Number. A number of contrast to take (by default - 2). Make sure to never write 1 here - it is the intercept}

\item{threshold_fc}{Number. A threshold to filter by FC. Doesn't do anything if do_fc_filtration=false}

\item{threshold_pval}{Number. A threshold to filter adjusted by p.val. Doesn't do anything if do_fc_filtration=false}

\item{method_string}{Character string. One of possible methods for adjustment for stats::p.adjust. Hommel by default}

\item{output_plot_path}{A path to a file. If not NA - a plot of mena-variance trend will be written there.}

\item{pre_filtration}{Boolean. Do we do FC filtration before p. adjustment? True by default.}

\item{do_fc_filtration}{Boolean. Do we do FC filtration at all? True by default.}
}
\description{
This will be a test run for adding roxygen comment to files
This is wrapper to run limma. Warning - it requires specifically formatted local ensembl annotation to run correctly
}
