% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_limma_drop_unestimable.r
\name{do_limma_drop_unestimable}
\alias{do_limma_drop_unestimable}
\title{A wrapper around limma that automatically reruns it while dropping unestimable variables from design matrix.}
\usage{
do_limma_drop_unestimable(
  design_matrix,
  output_file_result,
  output_file_degs,
  output_plot_path,
  ...
)
}
\arguments{
\item{design_matrix}{Numeric matrix. A design matrix, with smaples in row matching columns of count and and variables in columns. Should be generated through model.matrix from stats. Passes it into}

\item{output_file_result}{A path to a file. FC and p.val for chosen contrast for each expressed gene are going to be written in the file with that path}

\item{output_file_degs}{A path to a file. FC and p.val for only degs (if any) for each expressed gene are going to be written in the file with that path}

\item{output_plot_path}{A path to a file. If not NA - a plot of mena-variance trend will be written there.}

\item{...}{Other arguments you need to pass to do_limma}
}
\description{
A wrapper around limma that automatically reruns it while dropping unestimable variables from design matrix.
}
