% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_number_n_from_list.r
\name{element_number_n_from_list}
\alias{element_number_n_from_list}
\title{This function returns element number n from each entry in a 2d list.}
\usage{
element_number_n_from_list(x, n = 1)
}
\arguments{
\item{x}{2d list.}

\item{n}{Number of element from list to get.}
}
\value{
A vector of elements number n.
}
\description{
This function returns element number n from each entry in a 2d list.
}
