% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_column_and_reorder_by_colname.r
\name{get_column_and_reorder_by_colname}
\alias{get_column_and_reorder_by_colname}
\title{This function gets cols in order by inputing df and colnames.}
\usage{
get_column_and_reorder_by_colname(input_df, col_names_in_order)
}
\arguments{
\item{input_df}{Data frame to get calls from.}

\item{col_names_in_order}{Vector of colnames to get.}
}
\value{
Data frame with cols in order.
}
\description{
This function gets cols in order by inputing df and colnames.
}
