% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_to_nth_position.r
\name{input_to_nth_position}
\alias{input_to_nth_position}
\title{This function inputs a new column to the designated position in a df.}
\usage{
input_to_nth_position(
  data_frame_input,
  vector_input,
  position = NA,
  name_of_new_column = "new_column"
)
}
\arguments{
\item{data_frame_input}{Data frame to input column in it.}

\item{vector_input}{Column to add to the data frame.}

\item{position}{Where to put a new column. If NA - put it at the end.}

\item{name_of_new_column}{Name for a new column}
}
\value{
A data frame with a column added to a position.
}
\description{
This function inputs a new column to the designated position in a df.
}
