% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_column_unique_simple.r
\name{make_column_unique_simple}
\alias{make_column_unique_simple}
\title{Simple version of "make column unique" functions, but just combines unique values with ", " delimiter}
\usage{
make_column_unique_simple(x, colnum = 1, delim = ", ")
}
\arguments{
\item{x}{Data frame to process.}

\item{colnum}{Number of a column to make unique.}

\item{delim}{Delimiter to paste unique values of.}
}
\value{
A data frame with a designated column having only unique values.
}
\description{
Simple version of "make column unique" functions, but just combines unique values with ", " delimiter
}
