% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_id_unique_by_column.r
\name{make_id_unique_by_column}
\alias{make_id_unique_by_column}
\title{This function makes ID's in data frame unique, selecting columns by the rule, applied to leading column. Usually used for cases where a gene has multiple probeset on microarray, and you have to choose one to get fc and p.val.
For example, if there is a leading column with values 10, 100, and -199 for a given unique value in id col, min rule will end up taking the row with -199, as will extreme.}
\usage{
make_id_unique_by_column(
  input_frame,
  id_column,
  leading_column,
  method = "extreme"
)
}
\arguments{
\item{input_frame}{Data frame to process.}

\item{id_column}{Number of a column to make unique.}

\item{leading_column}{This column is used to make decision which of rows to take.}

\item{method}{The method of choosing a row (applied to a leading column).}
}
\value{
A data frame with a designated column having only unique values.
}
\description{
This function makes ID's in data frame unique, selecting columns by the rule, applied to leading column. Usually used for cases where a gene has multiple probeset on microarray, and you have to choose one to get fc and p.val.
For example, if there is a leading column with values 10, 100, and -199 for a given unique value in id col, min rule will end up taking the row with -199, as will extreme.
}
