% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_to_object_list_on_listrows.r
\name{match_to_object_list_on_listrows}
\alias{match_to_object_list_on_listrows}
\title{This function checks for presence on an object in the list for every element of a 2 d list, and returns a 1 d list matching in any object of the list is present in a 2d part of the list.}
\usage{
match_to_object_list_on_listrows(input_list, object_list)
}
\arguments{
\item{input_list}{List to be cheked.}

\item{object_list}{vector of element to check list elements against.}
}
\value{
a boolean vector of presence of any elements from a object list in a vector in each slot of a list.
}
\description{
This function checks for presence on an object in the list for every element of a 2 d list, and returns a 1 d list matching in any object of the list is present in a 2d part of the list.
}
