% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_comment_char.read.table.r
\name{multiple_comment_char.read.table}
\alias{multiple_comment_char.read.table}
\title{This function acts same as read table, but is able to handle comments with more than 1 char, like ## for parsing vcfs.
I didn't write it, I stole it from somwhere on stack overflow.}
\usage{
multiple_comment_char.read.table(file, comment.char = "##", ...)
}
\arguments{
\item{file}{input path for file to read}

\item{comment.char}{new comment char to process.}

\item{...}{Various parameters you would pass to a read.table.}
}
\value{
Table from  without multi-comment lines.
}
\description{
This function acts same as read table, but is able to handle comments with more than 1 char, like ## for parsing vcfs.
I didn't write it, I stole it from somwhere on stack overflow.
}
