% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_filtration_for_merge.r
\name{name_filtration_for_merge}
\alias{name_filtration_for_merge}
\title{This function firlters contents of a two rows (or cols) in two dfs, and preserve only rows (or cols) which contain common elements between the two.
Warning could accept not vectors from dfs, but will result in unpredictable outputs.}
\usage{
name_filtration_for_merge(df_1, df_2, row_1, row_2, byrow = TRUE)
}
\arguments{
\item{df_1}{First data frame}

\item{df_2}{Second data frame}

\item{row_1}{First row to be used for filtering.}

\item{row_2}{Second row to be used for filtering.}

\item{byrow}{Use rows or calls? Putting false there will make script work by columns instead.}
}
\value{
A list, with 1 being a first filtered df,2 being a second filtered df and 3 having a list of common elements by which filtration was performed.
}
\description{
This function firlters contents of a two rows (or cols) in two dfs, and preserve only rows (or cols) which contain common elements between the two.
Warning could accept not vectors from dfs, but will result in unpredictable outputs.
}
