% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_simplifier.r
\name{name_simplifier}
\alias{name_simplifier}
\title{These scripts simplify names by removing some symbols. Mostly for system path creation and naming headers. _lower and upper also put them in a lower or upper case.}
\usage{
name_simplifier(
  x,
  replacement_in = " ",
  replacement_out = "_",
  deletion = c("\\\\)", "\\\\(", "\%")
)
}
\arguments{
\item{x}{Vector of characters to replace stuff in.}

\item{replacement_in}{Characters to replace.}

\item{replacement_out}{Characcters to put in instead.}

\item{deletion}{Characters to delete}
}
\value{
Vector of characters with replacements.
}
\description{
These scripts simplify names by removing some symbols. Mostly for system path creation and naming headers. _lower and upper also put them in a lower or upper case.
}
