% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_the_insides_of_a_list_with_separator.r
\name{paste_the_insides_of_a_list_with_separator}
\alias{paste_the_insides_of_a_list_with_separator}
\title{This function converts 2d list into 1d vector pasted insides. Yes, probably you can also do this with lapply.}
\usage{
paste_the_insides_of_a_list_with_separator(input.list, input.sep = ", ")
}
\arguments{
\item{input.list}{List to go through and paste.}

\item{input.sep}{Separator to paste with.}
}
\value{
Vector of characters with pasted insides.
}
\description{
This function converts 2d list into 1d vector pasted insides. Yes, probably you can also do this with lapply.
}
