% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_by_list.r
\name{replace_by_list}
\alias{replace_by_list}
\title{This function (mostly useful in processing GEO experiment matrices) to replace iteratively all of the cetrain pattern with a different string.}
\usage{
replace_by_list(to_replace, replacement, input_vec, ...)
}
\arguments{
\item{to_replace}{Vector of stuff to replace.}

\item{replacement}{Variable to replace it with.}

\item{input_vec}{Vector to replace stuff in.}

\item{...}{Other parameters to pass to gsub.}
}
\value{
Vector with replaced contents.
}
\description{
This function (mostly useful in processing GEO experiment matrices) to replace iteratively all of the cetrain pattern with a different string.
}
