% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_russian_numeric_with_universal.r
\name{replace_russian_numeric_with_universal}
\alias{replace_russian_numeric_with_universal}
\title{This function replacecs russian numeric decimal "," with universal "." and saves output as numeric.}
\usage{
replace_russian_numeric_with_universal(x)
}
\arguments{
\item{x}{Vector to replace decimals in. Most likely it will be character vector.}
}
\value{
Numeric vector with replaced decimals.
}
\description{
This function replacecs russian numeric decimal "," with universal "." and saves output as numeric.
}
