% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_hash.r
\name{retrieve_hash}
\alias{retrieve_hash}
\title{This function input a number of hash keys one by one in specified order, and returns a vector with list with}
\usage{
retrieve_hash(x, hash_name, list_mode = FALSE)
}
\arguments{
\item{x}{Vector of keys.}

\item{hash_name}{Hash from hash package.}

\item{list_mode}{Should the return be a list? Could be usefull if retrieveing from hashes, created with create_hash_link_for_multiple_values_per_key.r}
}
\value{
Vector/2d list of values obtained from hash.
}
\description{
This function input a number of hash keys one by one in specified order, and returns a vector with list with
}
