% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowttest.r
\name{rowttest}
\alias{rowttest}
\title{This is rowttest function for use with "apply" on data frames.
It takes vector of values and position of experiments and controls in vector and returns a single p.val}
\usage{
rowttest(input, experiments, controls, paired = FALSE)
}
\arguments{
\item{input}{Numeric vector. Values to run t.test on}

\item{experiments}{Numeric vector. Position of experiment values in an input vector.}

\item{controls}{Numeric vector. Position of control values in an input vector.}

\item{paired}{Boolean. Should t-test be paired or unpaired?}
}
\value{
Vector of p.values per row.
}
\description{
This is rowttest function for use with "apply" on data frames.
It takes vector of values and position of experiments and controls in vector and returns a single p.val
}
