% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_first_part.r
\name{separate_first_part}
\alias{separate_first_part}
\title{Function that gets the first part of string, splat by pattern.
For example "why are you here" splat by this function with default delim would return "why"}
\usage{
separate_first_part(x, pattern = " ")
}
\arguments{
\item{x}{Input string to be splat.}

\item{pattern}{Patter to use in strsplit. Single space by default}
}
\description{
Function that gets the first part of string, splat by pattern.
For example "why are you here" splat by this function with default delim would return "why"
}
