% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_last_part.r
\name{separate_last_part}
\alias{separate_last_part}
\title{Function that gets the last part of string, splat by pattern.
For example "why are you here" splat by this function with default delim would return "here"}
\usage{
separate_last_part(x, pattern = " ")
}
\arguments{
\item{x}{Input string to be splat.}

\item{pattern}{Patter to use in strsplit. Single space by default}
}
\value{
Splat part.
}
\description{
Function that gets the last part of string, splat by pattern.
For example "why are you here" splat by this function with default delim would return "here"
}
