% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_model_factor_constructor.r
\name{simple_model_factor_constructor}
\alias{simple_model_factor_constructor}
\title{This function creates simple factors from more complex ones, like when you need to manually create several individual factors from a single multi value factor.}
\usage{
simple_model_factor_constructor(x, def = "0", pattern = "1")
}
\arguments{
\item{x}{Vector of values to create factors of.}

\item{def}{Character or number - what will be first level factor?}

\item{pattern}{Character or number - what will be second level factor?}
}
\value{
Factor, in whicn first level is NOT identical to pattern, and second is what is identical to pattern.
}
\description{
This function creates simple factors from more complex ones, like when you need to manually create several individual factors from a single multi value factor.
}
