% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_df.r
\name{summary_df}
\alias{summary_df}
\title{This function gives a short summary of columns and their unique contents.}
\usage{
summary_df(input_i_df, max_length = 30, exluded = "")
}
\arguments{
\item{input_i_df}{Df to summarize}

\item{max_length}{Max length of unique entries in a col.}

\item{exluded}{Vector of contents to exclude from summary.}
}
\value{
No return. Pastes the output to STDOUT.
}
\description{
This function gives a short summary of columns and their unique contents.
}
