% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_missing_rows_from_x_to_y.r
\name{add_missing_rows_from_x_to_y}
\alias{add_missing_rows_from_x_to_y}
\title{This function adds missing (by rownames) rows from frame x to frame y.
param x Data frame. Donor of rows.
param y Data frame. Recepient of rows.
return Data frame y with rows from data frame x added.}
\usage{
add_missing_rows_from_x_to_y(x, y, fill = 0)
}
\description{
This function adds missing (by rownames) rows from frame x to frame y.
param x Data frame. Donor of rows.
param y Data frame. Recepient of rows.
return Data frame y with rows from data frame x added.
}
