% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_data_frame_column_type.r
\name{clone_data_frame_column_type}
\alias{clone_data_frame_column_type}
\title{This function copies the types of columns from one data frame to another.}
\usage{
clone_data_frame_column_type(donor, acceptor)
}
\arguments{
\item{donor}{Data frame from which column types will be copied.}

\item{acceptor}{Data frame to which column types will be copied.}
}
\value{
Data frame donor with column types from acceptor.
}
\description{
This function copies the types of columns from one data frame to another.
}
